+++ NEWS.md
@@ -1,7 +1,7 @@
 # stevemisc 1.7.0
 
 - Add `charitable_contributions`.
-- Scoped helper verbs (the "at" functions) are gradually getting `.by` support and, with it, breaking their link to the superseded `_at()` functions from `{dplyr}`. This starts with `lag_at()`, though will gradually cover all the "at" functions. Now includes `diff_at()`.
+- Scoped helper verbs (the "at" functions) are gradually getting `.by` support and, with it, breaking their link to the superseded `_at()` functions from `{dplyr}`. 
 
 # stevemisc 1.6.0